face-recognition-service/requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.1
numpy==1.22.4
opencv-python==4.5.5.64
face-recognition==1.3.0
face-recognition-models==0.3.0
pillow==9.0.1
pyjwt==2.3.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
gunicorn==20.1.0 